requirements.txt
alembic==1.10.2
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
flake8==6.0.0
Flask==2.2.3
Flask-JSON==0.3.5
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
numpy==1.24.2
pandas==1.5.3
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.6
pyflakes==3.0.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.5.post1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
